separate shs from replication actor and move all networking and replication to separate modules
