+++ src/actors/jsonrpc/config.rs
@@ -2,5 +2,3 @@
 pub const JSONRPC_IP: &str = "127.0.0.1";
 // Define the default port used for the JSON-RPC server.
 pub const JSONRPC_PORT: u16 = 3030;
-
-
