+++ src/actors/mod.rs
@@ -1,7 +1,5 @@
-pub mod connection_manager;
 pub mod ctrlc;
 pub mod jsonrpc;
-pub mod lan_discovery;
-pub mod peer;
+pub mod network;
+pub mod replication;
 pub mod rpc;
-pub mod tcp_server;
